+++ source/game_sa/Enums/eTargetDoor.h
@@ -1,0 +1,9 @@
+#pragma once
+
+enum eTargetDoor {
+    TARGET_DOOR_FRONT_LEFT  = 0,
+    TARGET_DOOR_FRONT_RIGHT = 8,
+    TARGET_DOOR_REAR_RIGHT  = 9,
+    TARGET_DOOR_DRIVER      = 10,
+    TARGET_DOOR_REAR_LEFT   = 11
+};
