+++ source/game_sa/Tasks/TaskTypes/TaskComplexEnterCar.h
@@ -9,14 +9,7 @@
 #include "TaskComplex.h"
 #include "PathFind.h"
 #include "TaskUtilityLineUpPedWithCar.h"
-
-enum eTargetDoor {
-    TARGET_DOOR_FRONT_LEFT  = 0,
-    TARGET_DOOR_FRONT_RIGHT = 8,
-    TARGET_DOOR_REAR_RIGHT  = 9,
-    TARGET_DOOR_DRIVER      = 10,
-    TARGET_DOOR_REAR_LEFT   = 11
-};
+#include "Enums/eTargetDoor.h"
 
 class CTaskComplexEnterCar : public CTaskComplex {
 public:
