+++ source/game_sa/Tasks/TaskTypes/TaskComplexLeaveAnyCar.cpp
@@ -1,5 +1,8 @@
 #include "StdInc.h"
 #include "TaskComplexLeaveAnyCar.h"
+#include "TaskComplexLeaveCar.h"
+#include "TaskSimpleNone.h"
+#include "Enums/eTargetDoor.h"
 
 void CTaskComplexLeaveAnyCar::InjectHooks() {
     RH_ScopedClass(CTaskComplexLeaveAnyCar);
@@ -12,7 +15,7 @@ void CTaskComplexLeaveAnyCar::InjectHooks() {
     RH_ScopedInstall(Clone_Reversed, 0x4211C0);
     RH_ScopedInstall(GetTaskType_Reversed, 0x421240);
     RH_ScopedInstall(CreateNextSubTask_Reversed, 0x63BCD0);
-    //RH_ScopedInstall(CreateFirstSubTask_Reversed, 0x63BCE0);
+    RH_ScopedInstall(CreateFirstSubTask_Reversed, 0x63BCE0);
     //RH_ScopedInstall(ControlSubTask_Reversed, 0x63BDC0);
 }
 
@@ -31,7 +34,10 @@ CTask* CTaskComplexLeaveAnyCar::CreateNextSubTask(CPed* ped) {
 
 // 0x63BCE0
 CTask* CTaskComplexLeaveAnyCar::CreateFirstSubTask(CPed* ped) {
-    return plugin::CallMethodAndReturn<CTask*, 0x63BCE0, CTaskComplexLeaveAnyCar*, CPed*>(this, ped);
+    if (ped->IsInVehicle()) {
+        return new CTaskComplexLeaveCar{ ped->m_pVehicle, TARGET_DOOR_FRONT_LEFT, m_delayTime, m_sensibleLeaveCar, m_forceGetOut };
+    }
+    return new CTaskSimpleNone{};
 }
 
 // 0x63BDC0
